requirements.txt
alembic==1.10.3
autopep8==1.6.0
bcrypt==4.0.1
flask-admin==1.6.1
flask-sqlalchemy==3.0.3
matplotlib==3.6.3
pandas==1.5.3
pip-chill==1.0.3
requests==2.28.1
